+++ platforms/web/src/src/lib/useListeners/index.ts
@@ -1,0 +1,2 @@
+export { useListeners } from './useListeners';
+export { sendInputEvent } from './event';
